+++ components/Dashboard/Aside/Aside.tsx
@@ -1,7 +1,6 @@
-// components/Dashboard/Aside/Aside.tsx
 import { getAside } from "@/lib/services/aside/getAside";
-import { AsideClient } from "./AsideClient";
-import { AsideHeader } from "./ui/AsideHeader";
+import { AsideClient } from "@/components/dashboard/aside/AsideClient";
+import { AsideHeader } from "@/components/dashboard/aside/ui/AsideHeader";
 
 export async function Aside() {
   const aside = await getAside();